+++ src/main/java/org/fugerit/java/nhg/reflect/config/EntryHelper.java
@@ -1,5 +1,7 @@
 package org.fugerit.java.nhg.reflect.config;
 
+import java.util.Collections;
+import java.util.Comparator;
 import java.util.List;
 
 public class EntryHelper {
@@ -17,6 +19,18 @@ public class EntryHelper {
         EntryMethod initMethod = new EntryMethod( INIT_METHOD_NAME );
         initMethod.setParameterTypes( parametersTypes );
         entry.getMethods().add( initMethod );
+        return entry;
+    }
+
+    private static Comparator<EntryMethod> SORT_METHODS = new Comparator<EntryMethod>() {
+        @Override
+        public int compare(EntryMethod o1, EntryMethod o2) {
+            return (o1.getName()+o1.getParameterTypes().size()).compareTo( o2.getName()+o2.getParameterTypes().size() );
+        }
+    };
+
+    public static Entry fixedOrder( Entry entry ) {
+        Collections.sort( entry.getMethods(), SORT_METHODS );
         return entry;
     }