+++ CHANGELOG.md
@@ -7,11 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- fixedSort method in EntryHelper
+
 ## [1.1.0] - 2024-03-24
 
 ### Added
 
-- InitHelper for constructors creations
+- EntryHelper for constructors creations
 - GenerateReflectConfig API to generate reflect-config entry
 
 ### Changed